requirements.txt
langchain_community
langchain_openai
streamlit
langchain_core
python-dotenv
langserve
fastapi
uvicorn